Split out Nuxeo fetcher
